perf: add serialization benchmarks, refactor to helpers
